Do not crash when nullable actions is null
Actions was made nullable in d93e351f4095d30a436fbf6ec3e3dc224e7905a1 but that same commit introduced a non-null assertion operator here which means it is valid to create a `KeyboardActionsConfig` with a null `actions` but this will crash `KeyboardActions` on init.

This commit ensures that we check if there are actions on the newConfig and only attempt to iterate over them when they exist. By assigning newConfig.actions to a local variable it is implicitly unwrapped once the null check has occurred.